Add quick word wrap fix
If the window is too narrow then the whole thing disappears. This is not
good. Evidently (at least) Chrome doesn't do wrapping along with the
weird vertical centring thing, so it may be prudent to get the width of
the window (and maybe the font used somehow...) and calculate some value
of characters displayed before wrapping from that.

Basically this is a super quick and dirty fix, that should really be
fixed by restructuring the whole thing so there's no weird bug where
everything disappears.